+++ client/src/components/Stores/Pickup/TakenSlot.vue
@@ -47,7 +47,7 @@
             v-b-tooltip="$i18n('pickup.copyNumber')"
             variant="outline-primary"
             size="sm"
-            @click="copyIntoClipboard(phoneNumber)"
+            @click="copyToClipboard(phoneNumber)"
           >
             <i class="fas fa-fw fa-clone" :aria-label="$i18n('pickup.copyNumber')" />
           </b-button>
@@ -147,7 +147,6 @@
 
 <script>
 import Avatar from '@/components/Avatar/Avatar.vue'
-import { pulseSuccess } from '@/script'
 import PhoneNumbers from '@/helper/phone-numbers'
 import conversationStore from '@/stores/conversations'
 import DataUser from '@/stores/user'
@@ -155,9 +154,11 @@ import StoreData, { STORE_LOG_ACTION } from '@/stores/stores'
 
 import { v4 as uuidv4 } from 'uuid'
 import PickupsData from '@/stores/pickups'
+import CopyToClipboardMixin from '@/mixins/CopyToClipboardMixin'
 
 export default {
   components: { Avatar },
+  mixins: [CopyToClipboardMixin],
   props: {
     date: {
       type: Date,
@@ -307,13 +308,6 @@ export default {
 
       if (lastFetchTimestamp !== null) {
         return new Date(lastFetchTimestamp * MILLISECONDS_PER_SECOND)
-      }
-    },
-    copyIntoClipboard (text) {
-      if (navigator.clipboard) {
-        navigator.clipboard.writeText(text).then(() => {
-          pulseSuccess(this.$i18n('pickup.copiedNumber', { number: text }))
-        })
       }
     },
     openChat () {