+++ src/Modules/Report/ReportGateway.php
@@ -26,182 +26,6 @@ class ReportGateway extends BaseGateway
         );
     }
 
-    public function getFoodsaverBetriebe($fsId): array
-    {
-        $stm = '
-			SELECT 	b.id, b.name
-			FROM 	fs_betrieb_team t,
-					fs_betrieb b
-			WHERE 	t.betrieb_id = b.id
-			AND 	t.foodsaver_id = :foodsaver_id
-		';
-
-        return $this->db->fetchAll($stm, [':foodsaver_id' => (int)$fsId]);
-    }
-
-    public function delReport($id): void
-    {
-        $this->db->delete('fs_report', ['id' => (int)$id]);
-    }
-
-    public function confirmReport($id): void
-    {
-        $this->db->update('fs_report', ['committed' => 1], ['id' => $id]);
-    }
-
-    public function getReportedSavers(): array
-    {
-        return $this->db->fetchAll(
-            '
-			SELECT 	fs.name,
-					CONCAT(fs.nachname," (",COUNT(rp.foodsaver_id),")") AS nachname,
-					fs.photo,
-					fs.id,
-					fs.sleep_status,
-					COUNT(rp.foodsaver_id) AS count,
-					CONCAT("/?page=report&sub=foodsaver&id=",fs.id) AS `href`
-
-			FROM 	fs_foodsaver fs,
-					fs_report rp
-
-			WHERE 	rp.foodsaver_id = fs.id
-
-			GROUP 	BY rp.foodsaver_id
-
-			ORDER BY count DESC, fs.name
-		'
-        );
-    }
-
-    public function getReportStats(): array
-    {
-        $ret = $this->db->fetchAll(
-            '
-			SELECT 	`committed`, COUNT(`id`) as count
-			FROM 	fs_report
-			GROUP BY `committed`
-			ORDER BY `committed`
-		'
-        );
-        $new = 0;
-        $com = 0;
-        foreach ($ret as $r) {
-            if ($r['committed'] == 0) {
-                $new = $r['count'];
-            } else {
-                $com = $r['count'];
-            }
-        }
-
-        return [
-            'com' => $com,
-            'new' => $new
-        ];
-    }
-
-    public function getReportedSaver($id): ?array
-    {
-        $stm = '
-			SELECT 	`id`,
-					`name`,
-					`nachname`,
-					`photo`,
-					sleep_status
-
-			FROM 	`fs_foodsaver`
-
-			WHERE 	id = :id
-		';
-        if ($fs = $this->db->fetch($stm, [':id' => (int)$id])
-        ) {
-            $stm = '
-				SELECT
-					r.id,
-	            	r.`msg`,
-	            	r.`tvalue`,
-	            	r.`reporttype`,
-					r.`time`,
-					UNIX_TIMESTAMP(r.`time`) AS time_ts,
-
-					rp.id AS rp_id,
-					rp.name AS rp_name,
-					rp.nachname AS rp_nachname,
-					rp.photo AS rp_photo
-
-				FROM
-	            	`fs_report` r
-
-	         	LEFT JOIN
-	            	`fs_foodsaver` fs ON r.foodsaver_id = fs.id
-
-				LEFT JOIN
-	            	`fs_foodsaver` rp ON r.reporter_id = rp.id
-
-				WHERE
-					r.foodsaver_id = :id
-
-	          	ORDER BY
-					r.`time` DESC
-			';
-            $fs['reports'] = $this->db->fetchAll($stm, [':id' => (int)$id]);
-
-            return $fs;
-        }
-
-        return null;
-    }
-
-    public function getReport($id): ?array
-    {
-        $stm = '
-			SELECT
-				r.id,
-            	r.`msg`,
-            	r.`tvalue`,
-            	r.`reporttype`,
-				r.`time`,
-				r.committed,
-				r.betrieb_id,
-				UNIX_TIMESTAMP(r.`time`) AS time_ts,
-
-				fs.id AS fs_id,
-				fs.name AS fs_name,
-				fs.nachname AS fs_nachname,
-				fs.photo AS fs_photo,
-
-				rp.id AS rp_id,
-				rp.name AS rp_name,
-				rp.nachname AS rp_nachname,
-				rp.photo AS rp_photo
-
-			FROM
-            	`fs_report` r
-
-         	LEFT JOIN
-            	`fs_foodsaver` fs ON r.foodsaver_id = fs.id
-
-			LEFT JOIN
-            	`fs_foodsaver` rp ON r.reporter_id = rp.id
-
-			WHERE
-				r.`id` = :id
-		';
-        $report = $this->db->fetch($stm, [':id' => (int)$id]);
-        if (!$report) {
-            return null;
-        }
-
-        $stm = 'SELECT id, name FROM fs_betrieb WHERE id = :store_id';
-        if ($report['betrieb_id'] > 0 && $betrieb = $this->db->fetch(
-            $stm,
-            [':store_id' => (int)$report['betrieb_id']]
-        )) {
-            $report['betrieb'] = $betrieb;
-        }
-
-        return $report;
-    }
-
     private function reportSelectDbal(): QueryBuilder
     {
         return $this->db->builder()
@@ -221,12 +45,12 @@ class ReportGateway extends BaseGateway
                 'fs.nachname AS fs_nachname',
                 'fs.photo AS fs_photo',
                 'fs.email AS fs_email',
-                'fs.stadt AS fs_stadt',
 
                 'rp.id AS rp_id',
                 'rp.name AS rp_name',
                 'rp.nachname AS rp_nachname',
                 'rp.photo AS rp_photo',
+                'rp.email AS rp_email',
                 'b.name AS b_name')
             ->leftJoin('r', 'fs_foodsaver', 'fs', 'r.foodsaver_id = fs.id')
             ->leftJoin('r', 'fs_foodsaver', 'rp', 'r.reporter_id = rp.id')
@@ -235,25 +59,31 @@ class ReportGateway extends BaseGateway
             ->orderBy('r.time', 'DESC');
     }
 
-    public function getReportsByReporteeRegions($regions, ?array $excludeReportsWithUsers, ?array $onlyReportsWithUsers = null)
+    public function getReportsByUser(int $userId): array
     {
         $query = $this->reportSelectDbal();
+        $query->andWhere($query->expr()->eq('r.foodsaver_id', $userId));
 
-        if (is_array($regions)) {
-            if (!empty($regions)) {
-                // querybuilder ignores the where clause when $regions is empty
-                $query->andWhere($query->expr()->in('fs.bezirk_id', $regions));
-            } else {
-                return [];
-            }
-        }
+        // restrict access only to new reports to avoid social conflicts from old entries
+        $query->andWhere('time >= \'2021-01-01\'');
+
+        return $query->fetchAllAssociative();
+    }
+
+    public function getReportsByReporteeRegions(int $regionId, ?array $excludeReportsWithUsers, ?array $onlyReportsWithUsers = null)
+    {
+        $query = $this->reportSelectDbal();
+        $query->andWhere($query->expr()->eq('fs.bezirk_id', $regionId));
+
         if (!empty($excludeReportsWithUsers)) {
             $query->andWhere($query->expr()->notIn('r.reporter_id', $excludeReportsWithUsers));
             $query->andWhere($query->expr()->notIn('r.foodsaver_id', $excludeReportsWithUsers));
         }
         if (!empty($onlyReportsWithUsers)) {
-            $query->andWhere($query->expr()->In('r.reporter_id', $onlyReportsWithUsers));
-            $query->orWhere($query->expr()->In('r.foodsaver_id', $onlyReportsWithUsers));
+            $query->andWhere($query->expr()->or(
+                $query->expr()->In('r.reporter_id', $onlyReportsWithUsers),
+                $query->expr()->In('r.foodsaver_id', $onlyReportsWithUsers)
+            ));
         }
 
         // restrict access only to new reports to avoid social conflicts from old entries
@@ -262,27 +92,18 @@ class ReportGateway extends BaseGateway
         return $query->fetchAllAssociative();
     }
 
-    public function getReportsForRegionlessByReporterRegion($regions, $excludeReportsAboutUser = null)
+    public function getReportAffiliation(int $reportId): array
     {
-        $query = $this->reportSelectDbal();
-        $query->where('fs.bezirk_id = 0');
-        if (is_array($regions)) {
-            $query->andWhere($query->expr()->in('rp.bezirk_id', $regions));
-        }
-        if ($excludeReportsAboutUser !== null) {
-            $query->andWhere('fs.id != :exclude_id');
-            $query->setParameter('exclude_id', $excludeReportsAboutUser);
-        }
-
-        return $query->fetchAllAssociative();
+        return $this->db->fetch('SELECT
+                fs.id AS userId, fs.bezirk_id AS regionId
+            FROM fs_report r
+            JOIN fs_foodsaver fs ON fs.id = r.foodsaver_id
+            WHERE r.id = ?
+        ', [$reportId]);
     }
 
-    public function getReports($committed = '0'): array
+    public function deleteReport(int $reportId): void
     {
-        $query = $this->reportSelectDbal();
-        $query = $query->where('r.committed = :committed');
-        $query = $query->setParameter('committed', $committed);
-
-        return $query->fetchAllAssociative();
+        $this->db->delete('fs_report', ['id' => $reportId]);
     }
 }