+++ src/Permissions/ReportPermissions.php
@@ -6,6 +6,7 @@ use Foodsharing\Lib\Session;
 use Foodsharing\Modules\Core\DBConstants\Foodsaver\Role;
 use Foodsharing\Modules\Core\DBConstants\Region\WorkgroupFunction;
 use Foodsharing\Modules\Group\GroupFunctionGateway;
+use Foodsharing\Modules\Report\ReportGateway;
 use Foodsharing\Modules\Unit\CurrentUserUnitsInterface;
 
 class ReportPermissions
@@ -17,6 +18,7 @@ class ReportPermissions
         Session $session,
         GroupFunctionGateway $groupFunctionGateway,
         private readonly CurrentUserUnitsInterface $currentUserUnits,
+        private readonly ReportGateway $reportGateway,
     ) {
         $this->session = $session;
         $this->groupFunctionGateway = $groupFunctionGateway;
@@ -53,42 +55,27 @@ class ReportPermissions
         return false;
     }
 
-    public function mayAccessArbitrationReports(int $regionId): bool
+    public function mayAccessReportsForUser(int $userId): bool
     {
-        if ($this->session->mayRole(Role::ORGA)) {
-            return true;
+        if ($this->session->id() === $userId) {
+            return false;
         }
 
-        $arbitrationGroup = $this->groupFunctionGateway->getRegionFunctionGroupId($regionId, WorkgroupFunction::ARBITRATION);
-
-        if (!empty($arbitrationGroup)) {
-            if ($this->currentUserUnits->isAdminFor($arbitrationGroup)) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    public function mayAccessReportGroupReports(int $regionId): bool
-    {
-        if ($this->session->mayRole(Role::ORGA)) {
-            return true;
-        }
-
-        $reportGroup = $this->groupFunctionGateway->getRegionFunctionGroupId($regionId, WorkgroupFunction::REPORT);
-
-        if (!empty($reportGroup)) {
-            if ($this->currentUserUnits->isAdminFor($reportGroup)) {
-                return true;
-            }
-        }
-
-        return false;
+        return $this->session->mayRole(Role::ORGA);
     }
 
     public function mayHandleReports(): bool
     {
+        return $this->session->mayRole(Role::ORGA);
+    }
+
+    public function mayDeleteReport(int $reportId): bool
+    {
+        $report = $this->reportGateway->getReportAffiliation($reportId);
+        if ($this->session->id() === $report['userId']) {
+            return false;
+        }
+
         return $this->session->mayRole(Role::ORGA);
     }
 }