+++ tests/Api/ReportApiCest.php
@@ -56,7 +56,7 @@ class ReportApiCest
         $I->sendGET($I->apiReportListForRegion($this->region['id']));
         $I->seeResponseCodeIs(HttpCode::OK);
         codecept_debug($I->seeResponseCodeIs(HttpCode::OK));
-        codecept_debug($I->seeResponseContainsJson(['data' => ['fs_id' => $this->foodsaver['id'], 'rp_id' => $this->foodsaver['id']]]));
+        codecept_debug($I->seeResponseContainsJson(['fs_id' => $this->foodsaver['id'], 'rp_id' => $this->foodsaver['id']]));
     }
 
     public function cantSeeReportAboutFoodsaverInSubRegion(ApiTester $I): void
@@ -65,7 +65,7 @@ class ReportApiCest
         $I->addReport($this->foodsharer['id'], $this->subRegionFoodsaver['id']);
         $I->sendGET($I->apiReportListForRegion($this->region['id']));
         $I->seeResponseCodeIs(HttpCode::OK);
-        $I->cantSeeResponseContainsJson(['data' => ['fs_id' => $this->subRegionFoodsaver['id'], 'rp_id' => $this->foodsharer['id']]]);
+        $I->cantSeeResponseContainsJson(['fs_id' => $this->subRegionFoodsaver['id'], 'rp_id' => $this->foodsharer['id']]);
     }
 
     public function dontSeeReportAboutSelf(ApiTester $I): void
@@ -75,7 +75,7 @@ class ReportApiCest
         $I->addReport($this->subRegionFoodsaver['id'], $this->reportGroupAdmin['id']);
         $I->sendGET($I->apiReportListForRegion($this->region['id']));
         $I->seeResponseCodeIs(HttpCode::OK);
-        $I->dontSeeResponseContainsJson(['data' => ['fs_id' => $this->reportGroupAdmin['id']]]);
+        $I->dontSeeResponseContainsJson(['fs_id' => $this->reportGroupAdmin['id']]);
     }
 
     public function dontSeeReportAboutFoodsharerReporterNotInRegion(ApiTester $I): void
@@ -84,7 +84,7 @@ class ReportApiCest
         $I->addReport($this->reportGroupAdmin['id'], $this->foodsharer['id']);
         $I->sendGET($I->apiReportListForRegion($this->region['id']));
         $I->seeResponseCodeIs(HttpCode::OK);
-        $I->dontSeeResponseContainsJson(['data' => ['fs_id' => $this->foodsharer['id']]]);
+        $I->dontSeeResponseContainsJson(['fs_id' => $this->foodsharer['id']]);
     }
 
     public function ArbitrationAdminSeesReportAboutReportAdmin(ApiTester $I): void
@@ -93,7 +93,7 @@ class ReportApiCest
         $I->addReport($this->foodsharer['id'], $this->reportGroupAdmin['id']);
         $I->sendGET($I->apiReportListForRegion($this->region['id']));
         $I->seeResponseCodeIs(HttpCode::OK);
-        $I->seeResponseContainsJson(['data' => ['fs_id' => $this->reportGroupAdmin['id'], 'rp_id' => $this->foodsharer['id']]]);
+        $I->seeResponseContainsJson(['fs_id' => $this->reportGroupAdmin['id'], 'rp_id' => $this->foodsharer['id']]);
     }
 
     public function foodsaverCannotAccessReports(ApiTester $I): void